requirements.txt
attrs==21.2.0
aws-cdk-lib==2.0.0
beautifulsoup4==4.10.0
boto3==1.20.21
botocore==1.23.21
cattrs==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.9
constructs==10.0.10
idna==3.3
jmespath==0.10.0
jsii==1.47.0
Pillow==8.4.0
publication==0.0.3
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
typing_extensions==4.0.1
urllib3==1.26.7
